+++ packages/terra-form-textarea/src/terra-dev-site/doc/example/Textarea.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import Field from 'terra-form-field';
+import Textarea from 'terra-form-textarea';
+
+const DefaultTextArea = () => (
+  <Field label="Default Textarea">
+    <Textarea
+      size="small"
+      id="default"
+      ariaLabel="default label"
+    />
+  </Field>
+);
+
+export default DefaultTextArea;