+++ packages/terra-form-textarea/src/terra-dev-site/test/form-textarea/AutoResizableTextarea.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Textarea from '../../../Textarea';
 
-const textarea = () => <Textarea id="auto-resizable" cols="2" isAutoResizable defaultValue="Default Value" />;
+const textarea = () => <Textarea id="auto-resizable" cols="2" isAutoResizable defaultValue="Default Value" ariaLabel="label" />;
 
 export default textarea;