+++ packages/terra-form-textarea/src/terra-dev-site/test/form-textarea/SmallTextarea.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Textarea from '../../../Textarea';
 
-const textarea = () => <Textarea size="small" />;
+const textarea = () => <Textarea size="small" ariaLabel="label" />;
 
 export default textarea;